+++ src/main/resources/templates/member/findIdForm.html
@@ -23,7 +23,7 @@
                             <h2 class="title text-default">아이디 찾기</h2>
                         </div>
                         <div class="modal-body">
-                            <form action="/member/findId" method="post">
+                            <form action="/members/search/userid" method="post">
                                 <div class="form-group">
                                     <label for="phoneNumber" class="control-label">휴대 전화</label><br>
                                     <input type="text" class="form-control-phoneNumber" id="phoneNumber" name="phonenumber" placeholder="휴대전화 번호 입력" onchange="checkPhoneNumber_pattern('findAccount')">
