+++ src/main/resources/templates/member/findIdSucessPage.html
@@ -29,7 +29,7 @@
                         </div>
                         <div class="modal-footer-findAccountPage">
                             <a href="javascript:moveLogin()" type="button" class="btn btn-primary" style="width:47%">로그인 하러가기</a>&nbsp;
-                            <a href="/move/findPwdForm" type="button" class="btn btn-primary" style="width:47%">비밀번호 찾기</a>
+                            <a href="/findPwdForm" type="button" class="btn btn-primary" style="width:47%">비밀번호 찾기</a>
                         </div>
                     </div>
                 </div>
