+++ src/main/resources/templates/member/findPwdForm.html
@@ -23,7 +23,7 @@
               <h2 class="title text-default">비밀번호 찾기</h2>
             </div>
             <div class="modal-body">
-              <form action="/member/findPwd" method="post">
+              <form action="/members/search/pwd" method="post">
                 <div class="form-group">
                   <label for="userId" class="control-label">아이디(*)</label>
                   <input type="text" class="form-control" id="userId" name="userid" placeholder="아이디 입력" onchange="checkId_pattern()"><br>
