+++ src/main/resources/templates/member/login.html
@@ -23,7 +23,7 @@
                                 <h2 class="title text-default">로그인</h2>
                             </div>
                             <div class="modal-body">
-                                <form action="/member/login" method="post" id="loginForm">
+                                <form action="/members/login" method="post" id="loginForm">
                                     <div class="form-group">
                                         <label for="userId" class="control-label">아이디</label>
                                         <input type="text" class="form-control-userid" id="userId" name="userid" placeholder="아이디 입력"><br>
