+++ src/main/resources/templates/member/loginFail.html
@@ -12,11 +12,11 @@
 
         if(loginFail === "notExistAccount"){
             alert("존재하지 않는 아이디 입니다.");
-            location.href="/move/login";
+            location.href="/login";
         }
         else if(loginFail === "passwdNotCorrect"){
             alert("비밀번호가 일치하지 않습니다.");
-            location.href="/move/login";
+            location.href="/login";
         }
     </script>
 </head>
