+++ src/main/resources/templates/member/registerForm.html
@@ -23,7 +23,7 @@
                                 <h2 class="title text-default">회원 가입</h2>
                             </div>
                             <div class="modal-body">
-                                <form action="/member/register" method="post">
+                                <form action="/members/new" method="post">
                                     <div class="form-group">
                                         <label for="userId" class="control-label">아이디(*)</label><br>
                                         <input type="text" class="form-control-registerId" id="userId" name="userid" placeholder="아이디 입력" onchange="checkId_pattern()">
